requirements.txt
appdirs==1.4.0
astroid==1.4.9
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
click==6.7
coverage==4.3.4
coveralls==1.1
defusedxml==0.6.0
dj-rest-auth==1.0.7
Django==3.0.7
docopt==0.6.2
django-allauth==0.42.0
django-environ==0.4.5
django-extensions==2.2.9
djangorestframework==3.11.0
djangorestframework-gis==0.15
factory-boy==2.12.0
Faker==4.1.1
flake8==3.8.3
idna==2.9
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.6.1
nose==1.3.7
oauthlib==3.1.0
onesignal-sdk==2.0.0
packaging==16.8
pyparsing==2.1.10
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
wrapt==1.10.8
